% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_species.R
\name{match_species}
\alias{match_species}
\title{Match Plant Species}
\usage{
match_species(user_data)
}
\arguments{
\item{user_data}{A data frame containing plant species. It must have a column `物种`.}
}
\value{
A data frame containing matched species.
}
\description{
This function matches plant species names between a provided data frame and
an internal reference data set (`d1`).
}
\examples{
user_data <- data.frame(物种 = c("Species1", "Species2"))
matched_species <- match_species(user_data)
print(matched_species)
}
